requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2
dnspython==2.0.0
email-validator==1.1.1
Faker==5.8.0
flake8==3.8.3
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Session==0.3.2
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
six==1.15.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
